+++ CHANGELOG.md
@@ -12,7 +12,7 @@
   Pacemaker 2.1 ([rhbz#1985981])
 - All errors, warning and progress related output is now printed to stderr
   instead of stdout
-- Make roles `Promoted` and `Unpromoted` default
+- Make roles `Promoted` and `Unpromoted` default ([rhbz#1885293])
 - Make auto-deleting constraint default for `pcs resource move` command
   ([rhbz#1996062])
 
@@ -23,8 +23,14 @@
   `/remote/node_available`, `/remote/resource_status`
 - Pcsd no longer provides data in format used by web UI in pcs 0.9.142 and older
 
+### Deprecated
+- Legacy role names `Master` and `Slave` ([rhbz#1885293])
+- Option `--master` is deprecated and has been replaced by option `--promoted`
+  ([rhbz#1885293])
+
 [rhbz#1283805]: https://bugzilla.redhat.com/show_bug.cgi?id=1283805
 [rhbz#1881064]: https://bugzilla.redhat.com/show_bug.cgi?id=1881064
+[rhbz#1885293]: https://bugzilla.redhat.com/show_bug.cgi?id=1885293
 [rhbz#1985981]: https://bugzilla.redhat.com/show_bug.cgi?id=1985981
 [rhbz#1996062]: https://bugzilla.redhat.com/show_bug.cgi?id=1996062